+++ CHANGELOG.md
@@ -37,3 +37,4 @@ All notable changes to this project will be documented in this file.
 ## [0.0.7] - 2023-08-31
 
 - separate models.py into bagtiles.py and tiles.py
+- Added metod to calculate word value base